Create a frame to display the audio spectrogram
* The basic visualization infrastructure can later be used for
display purposes.
* The window event loop must run in the main thread. For this
to work, GameLogic instances must also stay in the main thread.
* In order not to block the main thread with audio processing,
audio stream and analysis is offloaded to a separate thread.
* In a future commit, the results of the audio analysis thread
will be sent to the main thread so that we can display the
audio spectrogram. This would also remove the print statements
from the analysis thread. After this change, the threads will
be designed such that
  * worker threads handle the core analysis logic
  * main thread only deals with user I/O and display